qweqw

+++ Block/OutTx.cpp
@@ -1,6 +1,6 @@
 #include "OutTx.h"
 
-OutTx::OutTx(std::string publicId_, int amount_) :
+OutTx::OutTx(std::string publicId_, float amount_) :
 	publicId(publicId_),
 	amount(amount_)
 {
@@ -11,7 +11,7 @@ std::string OutTx::getPublicId(void)
 	return publicId;
 }
 
-int OutTx::getAmount(void)
+float OutTx::getAmount(void)
 {
 	return amount;
 }
